% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitr_templ_html.R
\name{bitr_templ_html}
\alias{bitr_templ_html}
\title{bitR HTML template

Loads additional style and template file}
\usage{
bitr_templ_html(
  toc = TRUE,
  maintitle = NULL,
  subtitle = NULL,
  menu_url = "https://r2bit.com/",
  menu_name = "bitR",
  blank_target = TRUE,
  logo_img = NULL,
  path = ".",
  ...
)
}
\arguments{
\item{toc}{logical. should a table of contents be displayed?}

\item{maintitle}{character. title in header.}

\item{subtitle}{character. subtitle in header.}

\item{menu_url}{character. menu url in header.}

\item{menu_name}{character. munu url name in header.}

\item{blank_target}{logical. choice the behavior that links opening within 
the another page. if TURE then open another page (target = "_blank")}

\item{logo_img}{character. name of logo image file on top left.}

\item{path}{name of directory to generate report file. default is "."}

\item{...}{additional arguments provided to \code{html_document}}
}
\value{
An R Markdown output format.
}
\description{
bitR HTML template

Loads additional style and template file
}
\references{
https://github.com/UNHCR-WEB/unhcRstyle/blob/master/R/unhcr_templ_html.R
}
